// web/src/models/loveLanguage.ts
export type LoveLanguage = {
  id: number;
  first: "AF" | "SE" | "PE" | "TI" | "TO";
  second: "AF" | "SE" | "PE" | "TI" | "TO";
  third: "AF" | "SE" | "PE" | "TI" | "TO";
  fourth: "AF" | "SE" | "PE" | "TI" | "TO";
  fifth: "AF" | "SE" | "PE" | "TI" | "TO";
  firstScore: number;
  secondScore: number;
  thirdScore: number;
  fourthScore: number;
  fifthScore: number;
};